LayoutTests/js/codegen-peephole-locals-expected.txt

Tests whether peephole optimizations on bytecode properly deal with local registers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS if_less_test() is true
PASS if_else_less_test() is true
PASS conditional_less_test() is true
PASS logical_and_less_test() is true
PASS logical_or_less_test() is true
PASS do_while_less_test() is true
PASS while_less_test() is true
PASS for_less_test() is true
PASS successfullyParsed is true

TEST COMPLETE
